# optimize U-polynomials