@heltonmc 2022-07-20
Ok I've fixed the second order horner scheme to handle both odd and even tuples. This method works for both U-polynomial and asymptotic expansions.. It looks like once the tuple size gets large (>20) it becomes much more efficient.

On master
```julia
julia> @benchmark Bessels.Uk_poly_Jn20(1.2, 50.0, x, Float64) setup=(x=rand()+50.0)
BenchmarkTools.Trial: 10000 samples with 974 evaluations.
 Range (min … max):  71.816 ns … 109.990 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     72.268 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   72.424 ns ±   0.694 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

       ▁   ▄█▃▂▁▁                            ▁▁▁               ▁
  ▇▁▁▁▄█▅▃▅██████████▇▆▆▆▆▅▅▄▄▅▄▅▃▄▃▁▁▄▄▁▄▅▇███████▇▇▇▇▆▅▅▅▅▅▄ █
  71.8 ns       Histogram: log(frequency) by time      74.5 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels._α_αp_poly_30(10.0, x) setup=(x=rand()+100.0)
BenchmarkTools.Trial: 10000 samples with 796 evaluations.
 Range (min … max):  157.975 ns … 269.913 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     159.094 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   160.132 ns ±   7.359 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▄██▃▁▁  ▄▄▃   ▁                                             ▂
  ▅▇██████▇████████▇█▇▆▇▆▆▆▅▃▁▆▆▅▄▁▄▄▁▁▁▃▄▆▆▄▅▅▄▄▄▄▅▃▃▁▆▃▅▄▄▅▅▅ █
  158 ns        Histogram: log(frequency) by time        175 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

This PR:
```julia
julia> @benchmark Bessels.Uk_poly20(1.2, 50.0, x, Bessels.even_odd_poly_minus) setup=(x=rand()+50.0)
BenchmarkTools.Trial: 10000 samples with 993 evaluations.
 Range (min … max):  34.664 ns … 63.607 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     34.676 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   34.796 ns ±  0.823 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▁                                                          ▁
  ██▅▄▄▄▄▁▃▄▁▁▁▃▁▃▁▁▃▄▃▁▁▁▁▄▅▇██▆▅▆▆▆▄▅▅▄▄▃▃▃▄▃▁▁▄▄▁▃▄▅▄▆▆▅▆▆ █
  34.7 ns      Histogram: log(frequency) by time      37.9 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Bessels._α_αp_poly_30(10.0, x) setup=(x=rand()+100.0)
BenchmarkTools.Trial: 10000 samples with 939 evaluations.
 Range (min … max):  103.024 ns … 159.882 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     103.424 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   104.159 ns ±   4.013 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆█    ▄▂   ▁                                                  ▁
  ██▄▁▃▇██▇▇▅█▆▅▅▄▄▅▃▆▅▅▅▃▃▃▅▁▄▄▆█▇▇▅▄▄▅▁▃▄▄▄▄▁▁▃▃▃▁▄▄▃▃▄▃▃▃▄▅▅ █
  103 ns        Histogram: log(frequency) by time        122 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

So the U-poly is about 2x faster and the asymptotic about 1.5x faster. 

There is honestly still some inference issues that I don't think I will be able to figure out right now. 

Removing this branch...
 https://github.com/heltonmc/Bessels.jl/blob/d2dc8e582e4858fbbc7a1f85043e27f53ccd7631/src/U_polynomials.jl#L1-L7

Decreases the computational time by almost half... unsure what is going on there. Might just turn that into an issue to be solved to come back to after getting `bessely` implemented