@heltonmc 2022-07-21
Ok I think this is good to go now! I have changed the debye functions to output both `besselj` and `bessely` and optimized some of the polynomial evaluation this actually results in much less computation time for both values.

### On master to output just `besselj`
```julia
julia> @benchmark Bessels.besselj_debye(150.0, x) setup=(x=rand()+110.0)
BenchmarkTools.Trial: 10000 samples with 818 evaluations.
 Range (min … max):  151.460 ns … 190.930 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     152.249 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   152.530 ns ±   1.041 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

           ▂█▆▆▁                                                 
  ▁▁▁▁▂▅▅▇▆█████▆▄▄▄▃▃▁▂▂▁▁▁▁▁▁▁▁▁▁▁▁▂▂▃▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  151 ns           Histogram: frequency by time          155 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

### This PR to give both `besselj` and `bessely`
```julia
julia> @benchmark Bessels.besseljy_debye(150.0, x) setup=(x=rand()+110.0)
BenchmarkTools.Trial: 10000 samples with 956 evaluations.
 Range (min … max):  90.068 ns … 104.696 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     90.535 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   90.570 ns ±   0.562 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▄▇▇▄▁    ▄▇█▆▃                             ▁▂▁▁▁   ▁▂▂▂▁▁  ▂
  ▅███████▇▇▇███████▇▆▅▅▅▄▅▅█▆▆▄▁▁▃▃▆▅▆▁▁▃▁▃▃▆██████▇█████████ █
  90.1 ns       Histogram: log(frequency) by time      92.3 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```